Xavfsizlik + tuzatishlar: rate-limit, kalitlar yangilandi, narx/vazn fixes
XAVFSIZLIK:
- Rate limiting (slowapi): global 120/min, login 20/min, register 10/min,
  staff-request 5/min, select-role 20/min, carrier order 15/min (spam himoya)
- CORS bo'sh bo'lsa endi faqat Telegram domenlari (* emas)
- Ro'yxat himoyasi: barcodes/items max_length=200, tg_init_data max_length=4096
  (1 so'rovda 100k element/DB flood oldini oladi)

TUZATISHLAR:
- Admin panel narx: so'm -> dollar ($) — boshqa panellar bilan bir xil
- Kiloli mahsulot vazni: yaxlitlanmaydi, kasr kg saqlanadi (weight_kg float)

Eslatma: sizib chiqqan maxfiy kalitlar GitHub branch'lardan o'chirildi,
JWT/QR/ENCRYPTION/WEBHOOK kalitlari yangilandi (Render env yangilanishi kerak).

+++ backend/app/schemas/product.py
@@ -12,7 +12,7 @@ class ProductOut(BaseModel):
     category: str
     type: ProductType
     quantity: int
-    weight_kg: int
+    weight_kg: float
     unit_weight_kg: float | None = None
     box_weight_kg: float | None = None
     cargo_price: int
